sys-apps/hashboot: Fix incorrect source variable
When downloading the hashboot tarball, e.g. `hashboot-0.9.14.tar.gz` it
extracts to a folder without the version name, so `hashboot`.

This breaks when trying to `emerge hashboot` because the ebuild is
looking for the source code in
`/var/tmp/portage/sys-apps/hashboot-0.9.14/work/hashboot-0.9.14` when
it's actually located at
`/var/tmp/portage/sys-apps/hashboot-0.9.14/work/hashboot`.

In the ebuild that the `P` variable resolves to `hashboot-0.9.14`
and by default `S` and `P` are the same. `S` is what Portage uses to
figure out where the source is, which is why it's trying to find it in
`/var/tmp/portage/sys-apps/hashboot-0.9.14/work/hashboot-0.9.14`.

I managed to get it to build by editing the ebuild to set the `S`
variable as so: `S="${WORKDIR}/${PN}"`, which is what this commit
adds to the ebuild.