net-im/nheko: Don't rebuild if there's no update.
I unpacked all URIs from EGIT_REPO_URI. That is not how it was designed
so it broke a bit: The commit ID of coeurl (local) was compared to the
commit ID of nheko (remote).

The commit ID of coeurl was stored because it was used in the last call
to git-r3_src_unpack and it was compared to the commit ID of nheko
because it was the first in EGIT_REPO_URI.

I now unpack the dependencies first from a dedicated variable and then
unpack nheko like normal.